requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cleanup==6.0.0
django-extensions==3.2.1
django-tinymce==3.5.0
Pillow==9.3.0
psycopg2==2.9.5
sqlparse==0.4.3
